distro/fedora: Rename fedora pkg
